requirements.txt
python-telegram-bot==20.7
language-tool-python
requests
python-Levenshtein
tldextract
